[up:trimming-the-fat] move graph viz to utils, rm viz module
