+++ examples/mnist_feed_forward_nn.py
@@ -5,7 +5,7 @@ import nujo as nj
 import nujo.nn as nn
 import nujo.objective as obj
 import nujo.optim as optim
-from nujo.utils.viz import ComputationGraphPlotter
+from nujo.utils import ComputationGraphPlotter
 
 net = nn.Linear(28 * 28, 256) >> nn.Sigmoid() \
       >> nn.Linear(256, 128) >> nn.Sigmoid() \
