+++ nujo/utils/__init__.py
@@ -1,1 +1,7 @@
 ''' nujo utils '''
+
+from nujo.utils.computation_graph_plotter import ComputationGraphPlotter
+
+__all__ = [
+    'ComputationGraphPlotter',
+]
