--- nujo/utils/viz/__init__.py
@@ -1,5 +1,0 @@
-from nujo.utils.viz.computation_graph_plotter import ComputationGraphPlotter
-
-__all__ = [
-    'ComputationGraphPlotter',
-]
